Update mr_updater.sh
trying to best accommodate manjaro here since i noticed there was no checkupdates in their repo like arch has. instead manjaro seems to use pamac checkupdates.

+++ mr_updater.sh
@@ -741,46 +741,62 @@ stop_spinner() {
 # Function to update the system
 update_system() {
     case "$DISTRO_ID" in
-        "arch"|"manjaro"|"endeavouros")
+        "arch"|"endeavouros")
             echo -e "${ORANGE}==>> Checking 'pacman' packages to update...${NC}"
             # Use run_command to execute the pacman update
             # Run checkupdates and capture its output
         output=$(checkupdates 2>/dev/null)
 
-        # Check if updates are available based on output
-        if [[ -n "$output" ]]; then
-            echo -e "${ORANGE}  >> Updates found. Proceeding with system update...${NC}"
-            run_command "sudo pacman -Syyuu --noconfirm --needed --color=auto"
-        else
-            echo -e "${ORANGE}==>> Pacman packages are up-to-date. No updates required.${NC}"
-        fi
+            # Check if updates are available based on output
+            if [[ -n "$output" ]]; then
+                echo -e "${ORANGE}  >> Updates found. Proceeding with system update...${NC}"
+                run_command "sudo pacman -Syyuu --noconfirm --needed --color=auto"
+            else
+                echo -e "${ORANGE}  >> Pacman packages are up-to-date. No updates required.${NC}"
+            fi
+            ;;
+        manjaro)
+            echo -e "${ORANGE}==>> Checking for updates...${NC}"
+            if command -v pamac >/dev/null 2>&1; then
+                outputm=$(pamac checkupdates 2>/dev/null)
 
-            # Use the global AUR_PACKAGES variable to determine AUR updates
-            if [ -n "$AUR_PACKAGES" ]; then
-                echo -e "${ORANGE}==>> Inspecting yay cache...${NC}"
-                # Check yay cache exists
-                if [ -d "$HOME/.cache/yay" ]; then
-                    # Check if yay cache is empty
-                    if [ -z "$(find "$HOME/.cache/yay" -maxdepth 1 -type d | grep -v "^$HOME/.cache/yay$")" ]; then
-                        echo -e "${GREEN}  >> yay cache is clean${NC}"
-                    else
-                        # Collect directories to be cleaned
-                        mapfile -t yay_cache_dirs < <(find "$HOME/.cache/yay" -maxdepth 1 -type d | grep -v "^$HOME/.cache/yay$")
-                        
-                        if [ ${#yay_cache_dirs[@]} -gt 0 ]; then
-                            echo -e "${ORANGE}==>> Cleaning yay cache directories: ${NC}"
-                            for dir in "${yay_cache_dirs[@]}"; do
-                                printf "${WHITE}  - %s\n${NC}" "$(basename "$dir")"  # Use basename for each directory
-                                # Remove the directories
-                                rm -rf "$dir"
-                            done
-                        fi
-                    fi
+                # Check if the output contains the phrase "Your system is up to date."
+                if echo "$outputm" | grep -q "Your system is up to date."; then
+                    echo -e "${GREEN}  >> Your system is up-to-date. No updates required.${NC}"
                 else
-                    echo -e "${RED}!!! yay cache directory not found: $HOME/.cache/yay${NC}"
+                    echo -e "${ORANGE}  >> Updates found. Proceeding with system update...${NC}"
+                    run_command "sudo pacman -Syyuu --noconfirm --needed --color=auto"
                 fi
-                echo -e "${ORANGE}==>> Checking 'aur' packages to update...${NC}"
-                yay -Sua --norebuild --noredownload --removemake --answerclean A --noanswerdiff --noansweredit --noconfirm --cleanafter
+
+                # Use the global AUR_PACKAGES variable to determine AUR updates
+                if [ -n "$AUR_PACKAGES" ]; then
+                    echo -e "${ORANGE}==>> Inspecting yay cache...${NC}"
+                    # Check yay cache exists
+                    if [ -d "$HOME/.cache/yay" ]; then
+                        # Check if yay cache is empty
+                        if [ -z "$(find "$HOME/.cache/yay" -maxdepth 1 -type d | grep -v "^$HOME/.cache/yay$")" ]; then
+                            echo -e "${GREEN}  >> yay cache is clean${NC}"
+                        else
+                            # Collect directories to be cleaned
+                            mapfile -t yay_cache_dirs < <(find "$HOME/.cache/yay" -maxdepth 1 -type d | grep -v "^$HOME/.cache/yay$")
+
+                            if [ ${#yay_cache_dirs[@]} -gt 0 ]; then
+                                echo -e "${ORANGE}==>> Cleaning yay cache directories: ${NC}"
+                                for dir in "${yay_cache_dirs[@]}"; do
+                                    printf "${WHITE}  - %s\n${NC}" "$(basename "$dir")"  # Use basename for each directory
+                                    # Remove the directories
+                                    rm -rf "$dir"
+                                done
+                            fi
+                        fi
+                    else
+                        echo -e "${RED}!!! yay cache directory not found: $HOME/.cache/yay${NC}"
+                    fi
+                    echo -e "${ORANGE}==>> Checking 'aur' packages to update...${NC}"
+                    yay -Sua --norebuild --noredownload --removemake --answerclean A --noanswerdiff --noansweredit --noconfirm --cleanafter
+                fi
+            else
+                echo "Error: pamac was not found. Please install pamac."
             fi
             ;;
         "debian"|"ubuntu"|"linuxmint")
@@ -825,7 +841,7 @@ prompt_update() {
     while true; do
         # Use localized prompt
         get_system_language
-        read -rp "$(echo -e "${MAGENTA}$UPDATE_PROMPT")" answer
+        read -rp "$(echo -e "${MAGENTA}$UPDATE_PROMPT${NC}")" answer
         answer=$(echo "$answer" | tr '[:upper:]' '[:lower:]')
 
         if [[ -z "$answer" || "$answer" == "yes" || "$answer" == "y" ]]; then
